Time: 7 ms (10.28%), Space: 5.9 MB (86.36%) - LeetHub

+++ 1922-count-good-numbers/1922-count-good-numbers.cpp
@@ -1,6 +1,6 @@
 class Solution {
 public:
-    int countGoodNumbers( long n) {
+    int countGoodNumbers( long long n) {
         long long result = n%2 == 0 ? 1 : 5 ;
         long long x = 20;
         int MOD = 1000000007;
